support simple consumer supports subscribing to multiple topics
